reorganize structure per feature
