+++ src/main/java/com/oltruong/moneycontrol/fileupload/FileUploadController.java
@@ -1,12 +1,12 @@
-package com.oltruong.moneycontrol.controller;
+package com.oltruong.moneycontrol.fileupload;
 
 import com.oltruong.moneycontrol.exception.BadRequestException;
-import com.oltruong.moneycontrol.model.Operation;
-import com.oltruong.moneycontrol.model.Rule;
-import com.oltruong.moneycontrol.repository.OperationRepository;
-import com.oltruong.moneycontrol.repository.RuleRepository;
-import com.oltruong.moneycontrol.service.BankParser;
-import com.oltruong.moneycontrol.service.BudgetAnalyzer;
+import com.oltruong.moneycontrol.operation.Operation;
+import com.oltruong.moneycontrol.rule.Rule;
+import com.oltruong.moneycontrol.operation.OperationRepository;
+import com.oltruong.moneycontrol.rule.RuleRepository;
+import com.oltruong.moneycontrol.parser.BankParser;
+import com.oltruong.moneycontrol.analyzer.BudgetAnalyzer;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
