+++ src/main/java/com/oltruong/moneycontrol/operation/OperationRepository.java
@@ -1,9 +1,6 @@
-package com.oltruong.moneycontrol.repository;
-
-import com.oltruong.moneycontrol.model.Operation;
+package com.oltruong.moneycontrol.operation;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 import java.util.Optional;
