+++ src/main/java/com/oltruong/moneycontrol/parser/BankParser.java
@@ -1,6 +1,6 @@
-package com.oltruong.moneycontrol.service;
+package com.oltruong.moneycontrol.parser;
 
-import com.oltruong.moneycontrol.model.Operation;
+import com.oltruong.moneycontrol.operation.Operation;
 
 import java.util.List;
