+++ src/test/java/com/oltruong/moneycontrol/fileupload/FileUploadControllerTest.java
@@ -1,10 +1,10 @@
-package com.oltruong.moneycontrol.controller;
+package com.oltruong.moneycontrol.fileupload;
 
 import com.oltruong.moneycontrol.exception.BadRequestException;
-import com.oltruong.moneycontrol.repository.OperationRepository;
-import com.oltruong.moneycontrol.repository.RuleRepository;
-import com.oltruong.moneycontrol.service.BankParser;
-import com.oltruong.moneycontrol.service.BudgetAnalyzer;
+import com.oltruong.moneycontrol.operation.OperationRepository;
+import com.oltruong.moneycontrol.rule.RuleRepository;
+import com.oltruong.moneycontrol.parser.BankParser;
+import com.oltruong.moneycontrol.analyzer.BudgetAnalyzer;
 
 import org.junit.Before;
 import org.junit.Test;
